tasks

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -4,7 +4,7 @@ DELIMITER $$
 DROP PROCEDURE IF EXISTS ComputeAverageScoreForUser;
 CREATE PROCEDURE ComputeAverageScoreForUser (user_id INT)
 BEGIN
-    DECLARE new_score INT;
+    DECLARE new_score FLOAT;
 
     SELECT AVG(score) INTO new_score FROM corrections WHERE user_id = in_user_id;
     UPDATE users SET average_score = new_score WHERE id = in_user_id;
